@@ -1,6 +1,6 @@
 {
 	"name": "slash",
-	"version": "4.0.0",
+	"version": "5.1.0",
 	"description": "Convert Windows backslash paths to slash paths",
 	"license": "MIT",
 	"repository": "sindresorhus/slash",
@@ -12,8 +12,9 @@
 	},
 	"type": "module",
 	"exports": "./index.js",
+	"types": "./index.d.ts",
 	"engines": {
-		"node": ">=12"
+		"node": ">=14.16"
 	},
 	"scripts": {
 		"test": "xo && ava && tsd"
@@ -31,8 +32,8 @@
 		"convert"
 	],
 	"devDependencies": {
-		"ava": "^3.15.0",
-		"tsd": "^0.14.0",
-		"xo": "^0.38.2"
+		"ava": "^5.2.0",
+		"tsd": "^0.28.1",
+		"xo": "^0.54.2"
 	}
 }
