@@ -1,6 +1,6 @@
 {
 	"name": "globby",
-	"version": "13.2.2",
+	"version": "14.0.2",
 	"description": "User-friendly glob matching",
 	"license": "MIT",
 	"repository": "sindresorhus/globby",
@@ -11,9 +11,13 @@
 		"url": "https://sindresorhus.com"
 	},
 	"type": "module",
-	"exports": "./index.js",
+	"exports": {
+		"types": "./index.d.ts",
+		"default": "./index.js"
+	},
+	"sideEffects": false,
 	"engines": {
-		"node": "^12.20.0 || ^14.13.1 || >=16.0.0"
+		"node": ">=18"
 	},
 	"scripts": {
 		"bench": "npm update @globby/main-branch glob-stream fast-glob && node bench.js",
@@ -59,33 +63,27 @@
 		"git"
 	],
 	"dependencies": {
-		"dir-glob": "^3.0.1",
-		"fast-glob": "^3.3.0",
+		"@sindresorhus/merge-streams": "^2.1.0",
+		"fast-glob": "^3.3.2",
 		"ignore": "^5.2.4",
-		"merge2": "^1.4.1",
-		"slash": "^4.0.0"
+		"path-type": "^5.0.0",
+		"slash": "^5.1.0",
+		"unicorn-magic": "^0.1.0"
 	},
 	"devDependencies": {
 		"@globby/main-branch": "sindresorhus/globby#main",
-		"@types/node": "^20.3.3",
+		"@types/node": "^20.9.0",
 		"ava": "^5.3.1",
 		"benchmark": "2.1.4",
 		"glob-stream": "^8.0.0",
-		"rimraf": "^5.0.1",
-		"tempy": "^3.0.0",
-		"tsd": "^0.28.1",
-		"typescript": "^5.1.6",
-		"xo": "^0.54.2"
+		"tempy": "^3.1.0",
+		"tsd": "^0.30.4",
+		"xo": "^0.57.0"
 	},
 	"xo": {
 		"ignores": [
 			"fixtures"
-		],
-		"rules": {
-			"@typescript-eslint/consistent-type-definitions": "off",
-			"n/prefer-global/url": "off",
-			"@typescript-eslint/consistent-type-imports": "off"
-		}
+		]
 	},
 	"ava": {
 		"files": [
