@@ -1,19 +1,22 @@
 {
 	"name": "path-type",
-	"version": "4.0.0",
+	"version": "5.0.0",
 	"description": "Check if a path is a file, directory, or symlink",
 	"license": "MIT",
 	"repository": "sindresorhus/path-type",
+	"funding": "https://github.com/sponsors/sindresorhus",
 	"author": {
 		"name": "Sindre Sorhus",
 		"email": "sindresorhus@gmail.com",
-		"url": "sindresorhus.com"
+		"url": "https://sindresorhus.com"
 	},
+	"type": "module",
+	"exports": "./index.js",
 	"engines": {
-		"node": ">=8"
+		"node": ">=12"
 	},
 	"scripts": {
-		"test": "xo && nyc ava && tsd-check"
+		"test": "xo && nyc ava && tsd"
 	},
 	"files": [
 		"index.js",
@@ -26,7 +29,6 @@
 		"is",
 		"check",
 		"directory",
-		"dir",
 		"file",
 		"filepath",
 		"symlink",
@@ -37,9 +39,9 @@
 		"filesystem"
 	],
 	"devDependencies": {
-		"ava": "^1.3.1",
-		"nyc": "^13.3.0",
-		"tsd-check": "^0.3.0",
-		"xo": "^0.24.0"
+		"ava": "^3.15.0",
+		"nyc": "^15.1.0",
+		"tsd": "^0.14.0",
+		"xo": "^0.37.1"
 	}
 }
