@@ -1,10 +1,10 @@
 {
   "name": "postcss-cli",
-  "version": "10.1.0",
+  "version": "11.0.0",
   "description": "CLI for PostCSS",
   "type": "module",
   "engines": {
-    "node": ">=14"
+    "node": ">=18"
   },
   "bin": {
     "postcss": "./index.js"
@@ -22,9 +22,9 @@
     "dependency-graph": "^0.11.0",
     "fs-extra": "^11.0.0",
     "get-stdin": "^9.0.0",
-    "globby": "^13.0.0",
+    "globby": "^14.0.0",
     "picocolors": "^1.0.0",
-    "postcss-load-config": "^4.0.0",
+    "postcss-load-config": "^5.0.0",
     "postcss-reporter": "^7.0.0",
     "pretty-hrtime": "^1.0.3",
     "read-cache": "^1.0.0",
@@ -33,13 +33,13 @@
   },
   "devDependencies": {
     "ava": "^3.1.0",
-    "c8": "^7.9.0",
+    "c8": "^8.0.0",
     "coveralls": "^3.0.0",
-    "eslint": "^8.2.0",
-    "eslint-config-problems": "7.0.1",
+    "eslint": "^8.55.0",
+    "eslint-config-problems": "8.0.0",
     "postcss": "^8.0.4",
     "postcss-import": "^15.0.0",
-    "prettier": "~2.8.0",
+    "prettier": "~3.1.0",
     "sugarss": "^4.0.0",
     "uuid": "^9.0.0"
   },
