@@ -60,104 +109,135 @@
   },
   "homepage": "https://rollupjs.org/",
   "optionalDependencies": {
-    "fsevents": "~2.3.2"
+    "fsevents": "~2.3.2",
+    "@rollup/rollup-darwin-arm64": "4.30.1",
+    "@rollup/rollup-android-arm64": "4.30.1",
+    "@rollup/rollup-win32-arm64-msvc": "4.30.1",
+    "@rollup/rollup-freebsd-arm64": "4.30.1",
+    "@rollup/rollup-linux-arm64-gnu": "4.30.1",
+    "@rollup/rollup-linux-arm64-musl": "4.30.1",
+    "@rollup/rollup-android-arm-eabi": "4.30.1",
+    "@rollup/rollup-linux-arm-gnueabihf": "4.30.1",
+    "@rollup/rollup-linux-arm-musleabihf": "4.30.1",
+    "@rollup/rollup-win32-ia32-msvc": "4.30.1",
+    "@rollup/rollup-linux-loongarch64-gnu": "4.30.1",
+    "@rollup/rollup-linux-riscv64-gnu": "4.30.1",
+    "@rollup/rollup-linux-powerpc64le-gnu": "4.30.1",
+    "@rollup/rollup-linux-s390x-gnu": "4.30.1",
+    "@rollup/rollup-darwin-x64": "4.30.1",
+    "@rollup/rollup-win32-x64-msvc": "4.30.1",
+    "@rollup/rollup-freebsd-x64": "4.30.1",
+    "@rollup/rollup-linux-x64-gnu": "4.30.1",
+    "@rollup/rollup-linux-x64-musl": "4.30.1"
+  },
+  "dependencies": {
+    "@types/estree": "1.0.6"
   },
   "devDependenciesComments": {
-    "@rollup/plugin-typescript": "It appears that 11.1.3 breaks sourcemaps"
+    "core-js": "We only update manually as every update requires a snapshot update"
   },
   "devDependencies": {
-    "@codemirror/commands": "^6.2.5",
-    "@codemirror/lang-javascript": "^6.2.1",
-    "@codemirror/language": "^6.9.0",
-    "@codemirror/search": "^6.5.3",
-    "@codemirror/state": "^6.2.1",
-    "@codemirror/view": "^6.19.0",
-    "@jridgewell/sourcemap-codec": "^1.4.15",
-    "@mermaid-js/mermaid-cli": "^10.4.0",
-    "@rollup/plugin-alias": "^5.0.0",
-    "@rollup/plugin-buble": "^1.0.2",
-    "@rollup/plugin-commonjs": "^25.0.4",
-    "@rollup/plugin-json": "^6.0.0",
-    "@rollup/plugin-node-resolve": "^15.2.1",
-    "@rollup/plugin-replace": "^5.0.2",
-    "@rollup/plugin-terser": "^0.4.3",
-    "@rollup/plugin-typescript": "11.1.2",
-    "@rollup/pluginutils": "^5.0.4",
-    "@types/estree": "1.0.1",
-    "@types/mocha": "^10.0.1",
-    "@types/node": "~14.18.61",
-    "@types/yargs-parser": "^21.0.0",
-    "@typescript-eslint/eslint-plugin": "^6.7.2",
-    "@typescript-eslint/parser": "^6.7.2",
-    "@vue/eslint-config-prettier": "^8.0.0",
-    "@vue/eslint-config-typescript": "^12.0.0",
-    "acorn": "^8.10.0",
+    "@codemirror/commands": "^6.7.1",
+    "@codemirror/lang-javascript": "^6.2.2",
+    "@codemirror/language": "^6.10.8",
+    "@codemirror/search": "^6.5.8",
+    "@codemirror/state": "^6.5.0",
+    "@codemirror/view": "^6.36.1",
+    "@eslint/js": "^9.17.0",
+    "@inquirer/prompts": "^7.2.1",
+    "@jridgewell/sourcemap-codec": "^1.5.0",
+    "@mermaid-js/mermaid-cli": "^11.4.2",
+    "@napi-rs/cli": "^2.18.4",
+    "@rollup/plugin-alias": "^5.1.1",
+    "@rollup/plugin-buble": "^1.0.3",
+    "@rollup/plugin-commonjs": "^28.0.2",
+    "@rollup/plugin-json": "^6.1.0",
+    "@rollup/plugin-node-resolve": "^16.0.0",
+    "@rollup/plugin-replace": "^6.0.2",
+    "@rollup/plugin-terser": "^0.4.4",
+    "@rollup/plugin-typescript": "^12.1.2",
+    "@rollup/pluginutils": "^5.1.4",
+    "@shikijs/vitepress-twoslash": "^1.24.4",
+    "@types/mocha": "^10.0.10",
+    "@types/node": "^18.19.68",
+    "@types/semver": "^7.5.8",
+    "@types/yargs-parser": "^21.0.3",
+    "@vue/language-server": "^2.2.0",
+    "acorn": "^8.14.0",
     "acorn-import-assertions": "^1.9.0",
     "acorn-jsx": "^5.3.2",
-    "acorn-walk": "^8.2.0",
     "buble": "^0.20.0",
-    "builtin-modules": "^3.3.0",
-    "chokidar": "^3.5.3",
-    "colorette": "^2.0.20",
-    "concurrently": "^8.2.1",
-    "core-js": "^3.32.2",
+    "builtin-modules": "^4.0.0",
+    "chokidar": "^3.6.0",
+    "concurrently": "^9.1.0",
+    "core-js": "3.38.1",
+    "cross-env": "^7.0.3",
     "date-time": "^4.0.0",
     "es5-shim": "^4.6.7",
     "es6-shim": "^0.35.8",
-    "eslint": "^8.49.0",
-    "eslint-config-prettier": "^9.0.0",
-    "eslint-plugin-import": "^2.28.1",
-    "eslint-plugin-prettier": "^5.0.0",
-    "eslint-plugin-unicorn": "^48.0.1",
-    "eslint-plugin-vue": "^9.17.0",
+    "eslint": "^9.17.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-prettier": "^5.2.1",
+    "eslint-plugin-unicorn": "^56.0.1",
+    "eslint-plugin-vue": "^9.32.0",
     "fixturify": "^3.0.0",
     "flru": "^1.0.2",
-    "fs-extra": "^11.1.1",
+    "fs-extra": "^11.2.0",
     "github-api": "^3.4.0",
-    "hash.js": "^1.1.7",
-    "husky": "^8.0.3",
-    "inquirer": "^9.2.11",
-    "is-reference": "^3.0.2",
-    "lint-staged": "^14.0.1",
+    "globals": "^15.14.0",
+    "husky": "^9.1.7",
+    "is-reference": "^3.0.3",
+    "lint-staged": "^15.2.11",
     "locate-character": "^3.0.0",
-    "magic-string": "^0.30.3",
-    "mocha": "^10.2.0",
-    "nyc": "^15.1.0",
-    "pinia": "^2.1.6",
-    "prettier": "^3.0.3",
+    "magic-string": "^0.30.17",
+    "mocha": "^11.0.1",
+    "nodemon": "^3.1.9",
+    "npm-audit-resolver": "^3.0.0-RC.0",
+    "nyc": "^17.1.0",
+    "picocolors": "^1.1.1",
+    "pinia": "^2.3.0",
+    "prettier": "^3.4.2",
+    "prettier-plugin-organize-imports": "^4.1.0",
     "pretty-bytes": "^6.1.1",
-    "pretty-ms": "^8.0.0",
-    "requirejs": "^2.3.6",
-    "rollup": "^3.29.2",
-    "rollup-plugin-license": "^3.1.0",
+    "pretty-ms": "^9.2.0",
+    "requirejs": "^2.3.7",
+    "rollup": "^4.29.1",
+    "rollup-plugin-license": "^3.5.3",
     "rollup-plugin-string": "^3.0.0",
-    "rollup-plugin-thatworks": "^1.0.4",
-    "semver": "^7.5.4",
+    "semver": "^7.6.3",
     "shx": "^0.3.4",
     "signal-exit": "^4.1.0",
     "source-map": "^0.7.4",
     "source-map-support": "^0.5.21",
-    "systemjs": "^6.14.2",
-    "terser": "^5.19.4",
-    "tslib": "^2.6.2",
-    "typescript": "^5.2.2",
-    "vite": "^4.4.9",
-    "vitepress": "^1.0.0-rc.14",
-    "vue": "^3.3.4",
-    "weak-napi": "^2.0.2",
+    "systemjs": "^6.15.1",
+    "terser": "^5.37.0",
+    "tslib": "^2.8.1",
+    "typescript": "^5.7.2",
+    "typescript-eslint": "^8.18.2",
+    "vite": "^6.0.5",
+    "vitepress": "^1.5.0",
+    "vue": "^3.5.13",
+    "vue-tsc": "^2.2.0",
+    "wasm-pack": "^0.13.1",
     "yargs-parser": "^21.1.1"
   },
   "overrides": {
-    "semver": "^7.5.4"
+    "axios": "^1.7.9",
+    "semver": "^7.6.3",
+    "readable-stream": "npm:@built-in/readable-stream@1"
+  },
+  "comments": {
+    "vue-tsc": "This is necessary so that prettier-plugin-organize-imports works correctly in Vue templatges"
   },
   "files": [
+    "dist/*.node",
     "dist/**/*.js",
     "dist/*.d.ts",
     "dist/bin/rollup",
     "dist/es/package.json"
   ],
   "engines": {
-    "node": ">=14.18.0",
+    "node": ">=18.0.0",
     "npm": ">=8.0.0"
   },
   "exports": {
