@@ -1,17 +1,15 @@
-module.exports = (ctx) => ({
+import autoprefixer from 'autoprefixer';
+import cssnano from 'cssnano';
+
+export default {
     plugins: [
-        require('autoprefixer'),
-        ctx.env === 'production' ? require('cssnano')({
+        autoprefixer(),
+        process.env.NODE_ENV === 'production' ? cssnano({
             preset: ['default', {
                 discardComments: {
-                    removeAll: true,
-                },
-                normalizeWhitespace: true,
-                minifySelectors: true,
-                minifyParams: true,
-                reduceIdents: false,
-                reduceTransforms: false
+                    removeAll: true
+                }
             }]
         }) : false
     ].filter(Boolean)
-}); 
+}; 
